Feat: tailwind 적용 과정 + nextjs 리팩토링 수행중
