+++ next/.eslintrc.json
@@ -1,0 +1,3 @@
+{
+  "extends": ["next/babel", "next/core-web-vitals"]
+}
