+++ next/app/detail/page.tsx
