+++ next/app/layout.tsx
@@ -8,7 +8,11 @@ export const metadata = {
   description: "Generated by create next app",
 };
 
-export default function RootLayout({ children }) {
+type Props = {
+  children: React.ReactNode;
+};
+
+export default function RootLayout({ children }: Props) {
   return (
     <html lang="en">
       <body className={inter.className}>{children}</body>
