+++ next/package.json
@@ -9,7 +9,11 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@types/node": "^20.10.6",
     "bootstrap": "^5.3.2",
+    "eslint": "^8.56.0",
+    "eslint-config-next": "^14.0.4",
+    "framer-motion": "^10.17.4",
     "mongoose": "^8.0.3",
     "next": "14.0.4",
     "react": "^18",
